refactor([issue-1151]): split peer-sync, creative-director/importer, and story-builder schemas out of validation.js
validation.js drops from 2104 to ~1600 lines: the three domain groups move
to peerSyncValidation.js, creativeDirectorValidation.js (incl. the
Create-Suite importer block), and storyBuilderValidation.js, following the
existing brainValidation.js pattern (namespace exports in the lib barrel).
validation.js re-exports all of them so every existing deep import keeps
working — pinned by validationSplit.test.js (same-object identity + parse
smoke + primitives-stayed checks).

emptyToUndefined moved to zodCompat.js: ESM hoists validation.js's
'export * from' of the domain files, so they evaluate BEFORE its body —
importing it back from a domain file hits the TDZ (verified, documented at
the re-export site).

Refs #1151
